post_delete에서 method==POST일때만 실행되게 변경

+++ django/blog/views.py
@@ -75,6 +75,7 @@ def post_delete(request, pk):
     post = POST.objects.get(pk=pk)
     post.delete()
     """
-    post = Post.objects.get(pk=pk)
-    post.delete()
-    return redirect('post-list')
+    if request.method == 'POST':
+        post = Post.objects.get(pk=pk)
+        post.delete()
+        return redirect('post-list')
